requirements.txt
Django==6.0.3
djangorestframework==3.17.0
djangorestframework_simplejwt==5.5.1
django-cors-headers==4.9.0
django-filter==25.2
django-storages==1.14.6
dj-database-url==3.1.2
python-decouple==3.8
psycopg2-binary==2.9.11
boto3==1.42.71
Pillow==12.1.1
gunicorn==23.0.0
whitenoise==6.9.0
